fix(gateway): fix small
