+++ api/internal/gateway/user/v1/handler/auth_test.go
@@ -8,6 +8,7 @@ import (
 	"github.com/and-period/furumaru/api/internal/gateway/user/v1/response"
 	"github.com/and-period/furumaru/api/internal/user"
 	uentity "github.com/and-period/furumaru/api/internal/user/entity"
+	"github.com/and-period/furumaru/api/pkg/jst"
 	"github.com/golang/mock/gomock"
 )
 
@@ -244,6 +245,632 @@ func TestRefreshAuthToken(t *testing.T) {
 			t.Parallel()
 			const path = "/v1/auth/refresh-token"
 			req := newHTTPRequest(t, http.MethodPost, path, tt.req)
+			testHTTP(t, tt.setup, tt.expect, req)
+		})
+	}
+}
+
+func TestInitializeAuth(t *testing.T) {
+	t.Parallel()
+
+	tests := []struct {
+		name   string
+		setup  func(t *testing.T, mocks *mocks, ctrl *gomock.Controller)
+		req    *request.InitializeAuthRequest
+		expect *testResponse
+	}{
+		{
+			name: "success",
+			setup: func(t *testing.T, mocks *mocks, ctrl *gomock.Controller) {
+				in := &user.InitializeUserInput{
+					UserID:    idmock,
+					AccountID: "account-id",
+					Username:  "username",
+				}
+				mocks.user.EXPECT().InitializeUser(gomock.Any(), in).Return(nil)
+			},
+			req: &request.InitializeAuthRequest{
+				AccountID: "account-id",
+				Username:  "username",
+			},
+			expect: &testResponse{
+				code: http.StatusNoContent,
+			},
+		},
+		{
+			name: "failed to initialize user",
+			setup: func(t *testing.T, mocks *mocks, ctrl *gomock.Controller) {
+				in := &user.InitializeUserInput{
+					UserID:    idmock,
+					AccountID: "account-id",
+					Username:  "username",
+				}
+				mocks.user.EXPECT().InitializeUser(gomock.Any(), in).Return(errmock)
+			},
+			req: &request.InitializeAuthRequest{
+				AccountID: "account-id",
+				Username:  "username",
+			},
+			expect: &testResponse{
+				code: http.StatusInternalServerError,
+			},
+		},
+	}
+
+	for _, tt := range tests {
+		tt := tt
+		t.Run(tt.name, func(t *testing.T) {
+			t.Parallel()
+			const path = "/v1/auth/initialized"
+			req := newHTTPRequest(t, http.MethodPatch, path, tt.req)
+			testHTTP(t, tt.setup, tt.expect, req)
+		})
+	}
+}
+
+func TestUpdateUserAuth(t *testing.T) {
+	t.Parallel()
+
+	tests := []struct {
+		name   string
+		setup  func(t *testing.T, mocks *mocks, ctrl *gomock.Controller)
+		req    *request.UpdateAuthEmailRequest
+		expect *testResponse
+	}{
+		{
+			name: "success",
+			setup: func(t *testing.T, mocks *mocks, ctrl *gomock.Controller) {
+				in := &user.UpdateUserEmailInput{
+					AccessToken: tokenmock,
+					Email:       "test-user@and-period.jp",
+				}
+				mocks.user.EXPECT().UpdateUserEmail(gomock.Any(), in).Return(nil)
+			},
+			req: &request.UpdateAuthEmailRequest{
+				Email: "test-user@and-period.jp",
+			},
+			expect: &testResponse{
+				code: http.StatusNoContent,
+			},
+		},
+		{
+			name: "failed to update user email",
+			setup: func(t *testing.T, mocks *mocks, ctrl *gomock.Controller) {
+				in := &user.UpdateUserEmailInput{
+					AccessToken: tokenmock,
+					Email:       "test-user@and-period.jp",
+				}
+				mocks.user.EXPECT().UpdateUserEmail(gomock.Any(), in).Return(errmock)
+			},
+			req: &request.UpdateAuthEmailRequest{
+				Email: "test-user@and-period.jp",
+			},
+			expect: &testResponse{
+				code: http.StatusInternalServerError,
+			},
+		},
+	}
+
+	for _, tt := range tests {
+		tt := tt
+		t.Run(tt.name, func(t *testing.T) {
+			t.Parallel()
+			const path = "/v1/auth/email"
+			req := newHTTPRequest(t, http.MethodPatch, path, tt.req)
+			testHTTP(t, tt.setup, tt.expect, req)
+		})
+	}
+}
+
+func TestVerifyAuthEmail(t *testing.T) {
+	t.Parallel()
+
+	tests := []struct {
+		name   string
+		setup  func(t *testing.T, mocks *mocks, ctrl *gomock.Controller)
+		req    *request.VerifyAuthEmailRequest
+		expect *testResponse
+	}{
+		{
+			name: "success",
+			setup: func(t *testing.T, mocks *mocks, ctrl *gomock.Controller) {
+				in := &user.VerifyUserEmailInput{
+					AccessToken: tokenmock,
+					VerifyCode:  "123456",
+				}
+				mocks.user.EXPECT().VerifyUserEmail(gomock.Any(), in).Return(nil)
+			},
+			req: &request.VerifyAuthEmailRequest{
+				VerifyCode: "123456",
+			},
+			expect: &testResponse{
+				code: http.StatusNoContent,
+			},
+		},
+		{
+			name: "failed to veirify user email",
+			setup: func(t *testing.T, mocks *mocks, ctrl *gomock.Controller) {
+				in := &user.VerifyUserEmailInput{
+					AccessToken: tokenmock,
+					VerifyCode:  "123456",
+				}
+				mocks.user.EXPECT().VerifyUserEmail(gomock.Any(), in).Return(errmock)
+			},
+			req: &request.VerifyAuthEmailRequest{
+				VerifyCode: "123456",
+			},
+			expect: &testResponse{
+				code: http.StatusInternalServerError,
+			},
+		},
+	}
+
+	for _, tt := range tests {
+		tt := tt
+		t.Run(tt.name, func(t *testing.T) {
+			t.Parallel()
+			const path = "/v1/auth/email/verified"
+			req := newHTTPRequest(t, http.MethodPost, path, tt.req)
+			testHTTP(t, tt.setup, tt.expect, req)
+		})
+	}
+}
+
+func TestUpdateAuthPassword(t *testing.T) {
+	t.Parallel()
+
+	tests := []struct {
+		name   string
+		setup  func(t *testing.T, mocks *mocks, ctrl *gomock.Controller)
+		req    *request.UpdateAuthPasswordRequest
+		expect *testResponse
+	}{
+		{
+			name: "success",
+			setup: func(t *testing.T, mocks *mocks, ctrl *gomock.Controller) {
+				in := &user.UpdateUserPasswordInput{
+					AccessToken:          tokenmock,
+					OldPassword:          "!Qaz2wsx",
+					NewPassword:          "!Qaz3edc",
+					PasswordConfirmation: "!Qaz3edc",
+				}
+				mocks.user.EXPECT().UpdateUserPassword(gomock.Any(), in).Return(nil)
+			},
+			req: &request.UpdateAuthPasswordRequest{
+				OldPassword:          "!Qaz2wsx",
+				NewPassword:          "!Qaz3edc",
+				PasswordConfirmation: "!Qaz3edc",
+			},
+			expect: &testResponse{
+				code: http.StatusNoContent,
+			},
+		},
+		{
+			name: "failed to update user password",
+			setup: func(t *testing.T, mocks *mocks, ctrl *gomock.Controller) {
+				in := &user.UpdateUserPasswordInput{
+					AccessToken:          tokenmock,
+					OldPassword:          "!Qaz2wsx",
+					NewPassword:          "!Qaz3edc",
+					PasswordConfirmation: "!Qaz3edc",
+				}
+				mocks.user.EXPECT().UpdateUserPassword(gomock.Any(), in).Return(errmock)
+			},
+			req: &request.UpdateAuthPasswordRequest{
+				OldPassword:          "!Qaz2wsx",
+				NewPassword:          "!Qaz3edc",
+				PasswordConfirmation: "!Qaz3edc",
+			},
+			expect: &testResponse{
+				code: http.StatusInternalServerError,
+			},
+		},
+	}
+
+	for _, tt := range tests {
+		tt := tt
+		t.Run(tt.name, func(t *testing.T) {
+			t.Parallel()
+			const path = "/v1/auth/password"
+			req := newHTTPRequest(t, http.MethodPatch, path, tt.req)
+			testHTTP(t, tt.setup, tt.expect, req)
+		})
+	}
+}
+
+func TestForgotAuthPassword(t *testing.T) {
+	t.Parallel()
+
+	tests := []struct {
+		name   string
+		setup  func(t *testing.T, mocks *mocks, ctrl *gomock.Controller)
+		req    *request.ForgotAuthPasswordRequest
+		expect *testResponse
+	}{
+		{
+			name: "success",
+			setup: func(t *testing.T, mocks *mocks, ctrl *gomock.Controller) {
+				in := &user.ForgotUserPasswordInput{Email: "test-user@and-period.jp"}
+				mocks.user.EXPECT().ForgotUserPassword(gomock.Any(), in).Return(nil)
+			},
+			req: &request.ForgotAuthPasswordRequest{
+				Email: "test-user@and-period.jp",
+			},
+			expect: &testResponse{
+				code: http.StatusNoContent,
+			},
+		},
+		{
+			name: "failed to forget user password",
+			setup: func(t *testing.T, mocks *mocks, ctrl *gomock.Controller) {
+				in := &user.ForgotUserPasswordInput{Email: "test-user@and-period.jp"}
+				mocks.user.EXPECT().ForgotUserPassword(gomock.Any(), in).Return(errmock)
+			},
+			req: &request.ForgotAuthPasswordRequest{
+				Email: "test-user@and-period.jp",
+			},
+			expect: &testResponse{
+				code: http.StatusInternalServerError,
+			},
+		},
+	}
+
+	for _, tt := range tests {
+		tt := tt
+		t.Run(tt.name, func(t *testing.T) {
+			t.Parallel()
+			const path = "/v1/auth/forgot-password"
+			req := newHTTPRequest(t, http.MethodPost, path, tt.req)
+			testHTTP(t, tt.setup, tt.expect, req)
+		})
+	}
+}
+
+func TestResetAuthPassword(t *testing.T) {
+	t.Parallel()
+
+	tests := []struct {
+		name   string
+		setup  func(t *testing.T, mocks *mocks, ctrl *gomock.Controller)
+		req    *request.ResetAuthPasswordRequest
+		expect *testResponse
+	}{
+		{
+			name: "success",
+			setup: func(t *testing.T, mocks *mocks, ctrl *gomock.Controller) {
+				in := &user.VerifyUserPasswordInput{
+					Email:                "test-user@and-period.jp",
+					VerifyCode:           "123456",
+					NewPassword:          "!Qaz2wsx",
+					PasswordConfirmation: "!Qaz2wsx",
+				}
+				mocks.user.EXPECT().VerifyUserPassword(gomock.Any(), in).Return(nil)
+			},
+			req: &request.ResetAuthPasswordRequest{
+				Email:                "test-user@and-period.jp",
+				VerifyCode:           "123456",
+				Password:             "!Qaz2wsx",
+				PasswordConfirmation: "!Qaz2wsx",
+			},
+			expect: &testResponse{
+				code: http.StatusNoContent,
+			},
+		},
+		{
+			name: "failed to verify user password",
+			setup: func(t *testing.T, mocks *mocks, ctrl *gomock.Controller) {
+				in := &user.VerifyUserPasswordInput{
+					Email:                "test-user@and-period.jp",
+					VerifyCode:           "123456",
+					NewPassword:          "!Qaz2wsx",
+					PasswordConfirmation: "!Qaz2wsx",
+				}
+				mocks.user.EXPECT().VerifyUserPassword(gomock.Any(), in).Return(errmock)
+			},
+			req: &request.ResetAuthPasswordRequest{
+				Email:                "test-user@and-period.jp",
+				VerifyCode:           "123456",
+				Password:             "!Qaz2wsx",
+				PasswordConfirmation: "!Qaz2wsx",
+			},
+			expect: &testResponse{
+				code: http.StatusInternalServerError,
+			},
+		},
+	}
+
+	for _, tt := range tests {
+		tt := tt
+		t.Run(tt.name, func(t *testing.T) {
+			t.Parallel()
+			const path = "/v1/auth/forgot-password/verified"
+			req := newHTTPRequest(t, http.MethodPost, path, tt.req)
+			testHTTP(t, tt.setup, tt.expect, req)
+		})
+	}
+}
+
+func TestGetAuthUser(t *testing.T) {
+	t.Parallel()
+
+	now := jst.Date(2022, 1, 2, 18, 30, 0, 0)
+	u := &uentity.User{
+		ID:           "user-id",
+		ProviderType: uentity.ProviderTypeEmail,
+		Email:        "test@and-period.jp",
+		PhoneNumber:  "+819012345678",
+		Username:     "username",
+		ThumbnailURL: "https://and-period.jp/thumbnail.png",
+		CreatedAt:    now,
+		UpdatedAt:    now,
+		VerifiedAt:   now,
+	}
+
+	tests := []struct {
+		name   string
+		setup  func(t *testing.T, mocks *mocks, ctrl *gomock.Controller)
+		expect *testResponse
+	}{
+		{
+			name: "success",
+			setup: func(t *testing.T, mocks *mocks, ctrl *gomock.Controller) {
+				in := &user.GetUserInput{UserID: idmock}
+				mocks.user.EXPECT().GetUser(gomock.Any(), in).Return(u, nil)
+			},
+			expect: &testResponse{
+				code: http.StatusOK,
+				body: &response.AuthUserResponse{
+					AuthUser: &response.AuthUser{
+						ID:           "user-id",
+						Username:     "username",
+						ThumbnailURL: "https://and-period.jp/thumbnail.png",
+					},
+				},
+			},
+		},
+		{
+			name: "failed to get user",
+			setup: func(t *testing.T, mocks *mocks, ctrl *gomock.Controller) {
+				in := &user.GetUserInput{UserID: idmock}
+				mocks.user.EXPECT().GetUser(gomock.Any(), in).Return(nil, errmock)
+			},
+			expect: &testResponse{
+				code: http.StatusInternalServerError,
+			},
+		},
+	}
+
+	for _, tt := range tests {
+		tt := tt
+		t.Run(tt.name, func(t *testing.T) {
+			t.Parallel()
+			const path = "/v1/auth/user"
+			req := newHTTPRequest(t, http.MethodGet, path, nil)
+			testHTTP(t, tt.setup, tt.expect, req)
+		})
+	}
+}
+
+func TestCreateAuth(t *testing.T) {
+	t.Parallel()
+
+	tests := []struct {
+		name   string
+		setup  func(t *testing.T, mocks *mocks, ctrl *gomock.Controller)
+		req    *request.CreateAuthRequest
+		expect *testResponse
+	}{
+		{
+			name: "success",
+			setup: func(t *testing.T, mocks *mocks, ctrl *gomock.Controller) {
+				in := &user.CreateUserInput{
+					Email:                "test-user@and-period.jp",
+					PhoneNumber:          "+819012345678",
+					Password:             "!Qaz2wsx",
+					PasswordConfirmation: "!Qaz2wsx",
+				}
+				mocks.user.EXPECT().CreateUser(gomock.Any(), in).Return("user-id", nil)
+			},
+			req: &request.CreateAuthRequest{
+				Email:                "test-user@and-period.jp",
+				PhoneNumber:          "+819012345678",
+				Password:             "!Qaz2wsx",
+				PasswordConfirmation: "!Qaz2wsx",
+			},
+			expect: &testResponse{
+				code: http.StatusOK,
+				body: &response.CreateAuthResponse{
+					ID: "user-id",
+				},
+			},
+		},
+		{
+			name: "failed to create user",
+			setup: func(t *testing.T, mocks *mocks, ctrl *gomock.Controller) {
+				in := &user.CreateUserInput{
+					Email:                "test-user@and-period.jp",
+					PhoneNumber:          "+819012345678",
+					Password:             "!Qaz2wsx",
+					PasswordConfirmation: "!Qaz2wsx",
+				}
+				mocks.user.EXPECT().CreateUser(gomock.Any(), in).Return("", errmock)
+			},
+			req: &request.CreateAuthRequest{
+				Email:                "test-user@and-period.jp",
+				PhoneNumber:          "+819012345678",
+				Password:             "!Qaz2wsx",
+				PasswordConfirmation: "!Qaz2wsx",
+			},
+			expect: &testResponse{
+				code: http.StatusInternalServerError,
+			},
+		},
+	}
+
+	for _, tt := range tests {
+		tt := tt
+		t.Run(tt.name, func(t *testing.T) {
+			t.Parallel()
+			const path = "/v1/auth/user"
+			req := newHTTPRequest(t, http.MethodPost, path, tt.req)
+			testHTTP(t, tt.setup, tt.expect, req)
+		})
+	}
+}
+
+func TestVerifyAuth(t *testing.T) {
+	t.Parallel()
+
+	tests := []struct {
+		name   string
+		setup  func(t *testing.T, mocks *mocks, ctrl *gomock.Controller)
+		req    *request.VerifyAuthRequest
+		expect *testResponse
+	}{
+		{
+			name: "success",
+			setup: func(t *testing.T, mocks *mocks, ctrl *gomock.Controller) {
+				in := &user.VerifyUserInput{
+					UserID:     "user-id",
+					VerifyCode: "123456",
+				}
+				mocks.user.EXPECT().VerifyUser(gomock.Any(), in).Return(nil)
+			},
+			req: &request.VerifyAuthRequest{
+				ID:         "user-id",
+				VerifyCode: "123456",
+			},
+			expect: &testResponse{
+				code: http.StatusNoContent,
+			},
+		},
+		{
+			name: "failed to verify user",
+			setup: func(t *testing.T, mocks *mocks, ctrl *gomock.Controller) {
+				in := &user.VerifyUserInput{
+					UserID:     "user-id",
+					VerifyCode: "123456",
+				}
+				mocks.user.EXPECT().VerifyUser(gomock.Any(), in).Return(errmock)
+			},
+			req: &request.VerifyAuthRequest{
+				ID:         "user-id",
+				VerifyCode: "123456",
+			},
+			expect: &testResponse{
+				code: http.StatusInternalServerError,
+			},
+		},
+	}
+
+	for _, tt := range tests {
+		tt := tt
+		t.Run(tt.name, func(t *testing.T) {
+			t.Parallel()
+			const path = "/v1/auth/user/verified"
+			req := newHTTPRequest(t, http.MethodPost, path, tt.req)
+			testHTTP(t, tt.setup, tt.expect, req)
+		})
+	}
+}
+
+func TestCreateAuthWithOAuth(t *testing.T) {
+	t.Parallel()
+
+	now := jst.Date(2022, 1, 2, 18, 30, 0, 0)
+	u := &uentity.User{
+		ID:           "user-id",
+		ProviderType: uentity.ProviderTypeEmail,
+		Email:        "test@and-period.jp",
+		PhoneNumber:  "+819012345678",
+		Username:     "username",
+		ThumbnailURL: "https://and-period.jp/thumbnail.png",
+		CreatedAt:    now,
+		UpdatedAt:    now,
+		VerifiedAt:   now,
+	}
+
+	tests := []struct {
+		name   string
+		setup  func(t *testing.T, mocks *mocks, ctrl *gomock.Controller)
+		expect *testResponse
+	}{
+		{
+			name: "success",
+			setup: func(t *testing.T, mocks *mocks, ctrl *gomock.Controller) {
+				in := &user.CreateUserWithOAuthInput{AccessToken: tokenmock}
+				mocks.user.EXPECT().CreateUserWithOAuth(gomock.Any(), in).Return(u, nil)
+			},
+			expect: &testResponse{
+				code: http.StatusOK,
+				body: &response.AuthUserResponse{
+					AuthUser: &response.AuthUser{
+						ID:           "user-id",
+						Username:     "username",
+						ThumbnailURL: "https://and-period.jp/thumbnail.png",
+					},
+				},
+			},
+		},
+		{
+			name: "failed to create user",
+			setup: func(t *testing.T, mocks *mocks, ctrl *gomock.Controller) {
+				in := &user.CreateUserWithOAuthInput{AccessToken: tokenmock}
+				mocks.user.EXPECT().CreateUserWithOAuth(gomock.Any(), in).Return(nil, errmock)
+			},
+			expect: &testResponse{
+				code: http.StatusInternalServerError,
+			},
+		},
+	}
+
+	for _, tt := range tests {
+		tt := tt
+		t.Run(tt.name, func(t *testing.T) {
+			t.Parallel()
+			const path = "/v1/auth/user/oauth"
+			req := newHTTPRequest(t, http.MethodPost, path, nil)
+			testHTTP(t, tt.setup, tt.expect, req)
+		})
+	}
+}
+
+func TestDeleteAuth(t *testing.T) {
+	t.Parallel()
+
+	tests := []struct {
+		name   string
+		setup  func(t *testing.T, mocks *mocks, ctrl *gomock.Controller)
+		expect *testResponse
+	}{
+		{
+			name: "success",
+			setup: func(t *testing.T, mocks *mocks, ctrl *gomock.Controller) {
+				in := &user.DeleteUserInput{UserID: idmock}
+				mocks.user.EXPECT().DeleteUser(gomock.Any(), in).Return(nil)
+			},
+			expect: &testResponse{
+				code: http.StatusNoContent,
+			},
+		},
+		{
+			name: "failed to delete user",
+			setup: func(t *testing.T, mocks *mocks, ctrl *gomock.Controller) {
+				in := &user.DeleteUserInput{UserID: idmock}
+				mocks.user.EXPECT().DeleteUser(gomock.Any(), in).Return(errmock)
+			},
+			expect: &testResponse{
+				code: http.StatusInternalServerError,
+			},
+		},
+	}
+
+	for _, tt := range tests {
+		tt := tt
+		t.Run(tt.name, func(t *testing.T) {
+			t.Parallel()
+			const path = "/v1/auth/user"
+			req := newHTTPRequest(t, http.MethodDelete, path, nil)
 			testHTTP(t, tt.setup, tt.expect, req)
 		})
 	}
