+++ api/internal/gateway/user/v1/response/auth.go
@@ -9,6 +9,21 @@ type Auth struct {
 	TokenType    string `json:"tokenType"`    // トークン種別
 }
 
+// AuthUser - 認証中ユーザー情報
+type AuthUser struct {
+	ID           string `json:"id"`           // ユーザーID
+	Username     string `json:"username"`     // ユーザー名 (表示名)
+	ThumbnailURL string `json:"thumbnailUrl"` // サムネイルURL
+}
+
 type AuthResponse struct {
 	*Auth
+}
+
+type AuthUserResponse struct {
+	*AuthUser
+}
+
+type CreateAuthResponse struct {
+	ID string `json:"id"` // ユーザーID
 }
