--- api/internal/gateway/user/v1/response/user.go
@@ -1,12 +1,0 @@
-package response
-
-type CreateUserResponse struct {
-	ID string `json:"id"` // ユーザーID
-}
-
-type UserMeResponse struct {
-	ID           string `json:"id"`           // ユーザーID
-	Email        string `json:"email"`        // メールアドレス
-	PhoneNumber  string `json:"phoneNumber"`  // 電話番号
-	ThumbnailURL string `json:"thumbnailUrl"` // サムネイルURL
-}
